Refactor opinion groups' selected cluster key
