+++ services/agora/src/utils/component/analysis/analysisTypes.ts
@@ -1,0 +1,3 @@
+import { PolisKey } from "src/shared/types/zod";
+
+export type SelectedClusterKeyType = PolisKey | "all";
